+++ src/components/Util/variants.js
@@ -1,0 +1,7 @@
+module.exports = {
+  all: color => [
+    `ripple-${color}-500`,
+    `ripple-${color}`,
+    `hover:bg-${color}-trans`
+  ]
+};